requirements.txt
openai==0.28.1
pyyaml==6.0.1
jinja2==3.1.2
